swap group/split nav keybindings; swank launches slime automatically
this commit swaps some of my keybindings, so that f and b move forward and
backward frames (i.e. splits) and n and p move to the next and previous
group (i.e. virtual desktops; workspaces). this better suits my geometric
intuitions about the layout of these constructs - frames are organized along
the horizontal axis, and groups along the vertical (note that using f and b to
move horizontally and n and p to move vertically come from emacs).

also, this expands my command SWANK to open slime with emacsclient and connect
to the newly opened swank server. i would like one day to expand the command
further, so that it only creates a swank server if one is not already running,
and switches to the slime repl otherwise, similar to STUMPWM:RUN-OR-RAISE.